Convert to typescript

+++ src/commands/statusclear.ts
@@ -1,6 +1,6 @@
 /*
 discord-gamestatus: Game server monitoring via discord API
-Copyright (C) 2019-2020 Douile
+Copyright (C) 2019-2021 Douile
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -13,11 +13,18 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
-const { isAdmin } = require('../checks.js');
-const { channelFirstArg } = require('../utils.js');
+import { TextChannel } from "discord.js-light";
 
-const call = async function(message, args) {
-  let response = await message.channel.send('Clearing status updates...');
+import Message from "../structs/Message";
+import { isAdmin } from "../checks";
+import { channelFirstArg } from "../utils";
+
+export const name = "statusclear";
+export const check = isAdmin;
+export const help = "Clear all status messages from the channel";
+
+export async function call(message: Message, args: string[]) {
+  let response = await message.channel.send("Clearing status updates...");
   let channel;
   try {
     channel = await channelFirstArg(message, args);
@@ -25,23 +32,26 @@ const call = async function(message, args) {
     return;
   }
 
-  const updates = await message.client.updateCache.get({ channel: channel.id });
+  if (!channel || !(channel instanceof TextChannel)) return;
+
+  const updates = await message.client.updateCache.get({
+    channel: channel.id,
+    guild: channel.guild.id
+  });
   for (let update of updates) {
     const msg = await update.getMessage(message.client);
     if (msg) {
       try {
         await msg.delete();
-      } catch(e) {
+      } catch (e) {
         // DO NOTHING
       }
     }
   }
 
-  const count = await message.client.updateCache.delete({guild: channel.guild.id, channel: channel.id});
+  const count = await message.client.updateCache.delete({
+    guild: channel?.guild?.id,
+    channel: channel?.id
+  });
   await response.edit(`${count} Status updates have been cleared`);
 }
-
-exports.name = 'statusclear';
-exports.call = call;
-exports.check = isAdmin;
-exports.help = 'Clear all status messages from the channel';
